:recycle: refactor: 상수분리 및 연결, 문서화추가
