+++ src/domain/models/SpecialSaleEvent.js
@@ -1,4 +1,5 @@
-import { EVENT_PLANNER, SALE_PRIZE, EMPTY_COUNT } from '../../constants/constant.js';
+import { EMPTY_COUNT } from '../../constants/constant.js';
+import { EVENT_PLANNER, SALE_PRIZE } from '../../constants/conditions.js';
 
 class SpecialSaleEvent {
   #specialSalePrice;
